Logistic Regression
The aim of this project is to come up with principal components that help fit a regression model that can
ultimately be used for prediction. The point of this undertaking is to clarify the fluctuation in
autonomous factors or variables that determine an individual’s Income. In addition, a machine learning
approach will also be used to make predictions.